Change from int64_t to size_t in NStep agents

+++ reinforcement_learning/backend/include/agents/tabular/temporal_difference/n_step/n_step_expected_sarsa.hpp
@@ -8,7 +8,7 @@ namespace agents {
 
 class NStepExpectedSarsa : public NStep {
    public:
-    NStepExpectedSarsa(size_t states, size_t actions, float alpha, float epsilon, float gamma, int64_t n_step, float epsilon_decay = 1);
+    NStepExpectedSarsa(size_t states, size_t actions, float alpha, float epsilon, float gamma, size_t n_step, float epsilon_decay = 1);
     NStepExpectedSarsa(const NStepExpectedSarsa &other);
     NStepExpectedSarsa &operator=(const NStepExpectedSarsa &other);
     virtual ~NStepExpectedSarsa();
