+++ reinforcement_learning/backend/include/agents/tabular/temporal_difference/n_step/n_step_sarsa.hpp
@@ -8,7 +8,7 @@ namespace agents {
 
 class NStepSarsa : public NStep {
    public:
-    NStepSarsa(size_t states, size_t actions, float alpha, float epsilon, float gamma, int64_t n_step, float epsilon_decay = 1);
+    NStepSarsa(size_t states, size_t actions, float alpha, float epsilon, float gamma, size_t n_step, float epsilon_decay = 1);
     NStepSarsa(const NStepSarsa &other);
     NStepSarsa &operator=(const NStepSarsa &other);
     virtual ~NStepSarsa();
